requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-countries==5.5
django-heroku==0.3.1
django-language-field==0.0.3
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.4
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
